EcotoxSystems.jl

EcotoxSystems.defaultparams — Constant

Default parameter object

EcotoxSystems.default_ODE! — Method

DEB-ODE model with arbitrary number of stressors, assuming IA to compute combined effects.

EcotoxSystems.default_individual_ODE! — Method

Individual-level part of the DEB-ODE model with arbitrary number of stressors, assuming IA to compute combined effects.

EcotoxSystems.DEBkiss! — Method

DEBkiss!(du, u, p, t)::Nothing

Dynamics of DEBkiss model with maturity and explicit simulation of resource dynamics.

The density of structure is ignored, and instead S^(2/3) is applied for surface-area scaling. This affects the dimension of dI_max, but has no effect on the model dynamics.

If model output is to be compared to length data, a statistical weight-lenght relationship has to be applied to the model output.

EcotoxSystems.IBM_simulator — Method

IBM_simulator(
    p::ComponentVector; 
    dt = 1/24, 
    saveat = 1,
    showinfo::Number = Inf
    )::DataFrame

Simulate the individual-based version of the default model.

import EcotoxSystems as DEB
p = DEB.params()
sim = DEB.IBMsimulator(p)

args

p: The parameter collection with defined global and species parameters.

kwargs

dt: Length of a timestep in the model (unit according to chosen unit of rate parameters)

saveat: Time interval at which to record output

showinfo: Time interval at which to print an update. Nothing will be printed if showinfo == Inf (the default).

EcotoxSystems.ODE_simulator — Method

simulator(
    p::ComponentVector; 
    alg = Tsit5(),
    saveat = 1,
    reltol = 1e-6,
    model = default_ODE!,
    callbacks = DEBODE_callbacks,
    returntype::ReturnType = dataframe,
    kwargs...
)::DataFrame

Run the model as ODE system.

Example:

import EcotoxSystems as DEB
p = DEB.params() # loads the default parameters
sim = DEB.ODE_simulator(p)

EcotoxSystems.default_TKTD! — Method

function default_TKTD!(du, u, p, t)::Nothing

Mixture-TKTD for an arbitrary number of stressors, assuming Independent Action.

EcotoxSystems.calc_SL_max — Method

calc_SL_max(spc::ComponentVector)::Float64

Calculate maximum structural length slmax [m^(1/3)]

EcotoxSystems.calc_S_max — Method

calc_S_max(spc::ComponentVector)::Float64

Calculate maximum structural mass smax [m]

EcotoxSystems.clipneg — Method

Clip negative values at 0 as a continuous function, using sig.

EcotoxSystems.default_individual_rules! — Method

default_individual_rules(a::AbstractDEBIndividual, m::AbstractIBM)::Nothing

Defines the default rule-based portion for DEBIndividuals. <br>

The event functions which are used as callbacks during ODE solving are here re-used to apply rules for life stage transitions. A crude rule for starvation mortality is implemented, applying a dependency between starvation mortality and the scaled functional response.

Reproduction is assumed to occur in fixed time intervals, according to spc.tau_R.

EcotoxSystems.exposure — Method

exposure(
    simulator::Function, 
    p::ComponentVector, 
    C_Wmat::Matrix{R}
) where R <: Real

Simulate exposure to an arbitrary number of stressors over a Matrix of constant exposure concentrations C_Wmat.

The exposure matrix columns are stressors, the rows are treatments. That means, to simulate a single-stressor experiment, do

C_Wmat = [0.; 1.; 2;]

, creating a n x 1 matrix with exposure concentrations 0, 1 and 2.

In contrast, a single treatment with multiple stressors would be defined as

C_Wmat = [0 1 2;]

, creating a 1 x n matrix. Here, we would have three stressors with the simultaneous exposure concentrations 0,1,2. <br>

Thus, defining four treatments for two stressors looks like this:

C_Wmat = [0.0 0.0; 0.0 0.5; 1.0 0.0; 0.5 1.0]

This exposure matrix corresponds to a ray design with constant exposure ratios.

EcotoxSystems.generate_individual_params — Method

Generate individual-specific parameter set from species-specific parameter set. If a parameter entry is a distribution, a random sample is taken.

This also works for Vectors of distributions.

EcotoxSystems.individual_step! — Method

individual_step!(a::Agent, m::Model)::Nothing

The individual step follows a generic pattern:

First the ODE-portion of the model is executed, and the corresponding state variables are updated using the Euler scheme.

Then the rule-based portion of the model is executed. These are all the functions which cannot / should not be expressed as part of an ODE. At the minimum, this will include life stage transitions, reproduction and death of individuals.

For a spatially explicit model, movement should also most likely be part of the rule-based portion, as well as functions which require direct information exchange between individuals.

EcotoxSystems.initialize_statevars — Method

initialize_statevars(p::::ComponentVector, pindt::ComponentVector{Float64})::ComponentArray

For initialization of ODE simulator, initialize the component vector of state variables, u, based on common oaraeter collection p.

EcotoxSystems.k_J! — Method

k_J!(spc::ComponentVector)::Nothing

Set the maturity maintenance rate constant, assuming that the cumulative investment into maturity maintenance equals the cumulative investment into somatic maintenance (cf. DEBkiss book by Tjalling Jager).

EcotoxSystems.relative_response — Method

relative_response(
    sim::D, 
    response_vars::Vector{Symbol},
    treatment_var::Symbol; 
    groupby_vars::Vector{Symbol} = Symbol[],
    identify_control = minimum
    ) where D <: AbstractDataFrame

Calculate relative responses.

args:

sim::AbstractDataFrame: results

response_vars::Vector{Symbol}: response variables for which to calculate the relative responses

treatment_var::Symbol: Column indicating the treatment. Column values can be numerical or categorical, but identify_control kwarg has to be specified in the latter case

kwargs:

groupby_vars::Vector{Symbol}: relative response will be conditioned on these variables (e.g. time, separate experiments...). Empty by default.

identify_control: function used to identify reference values from treatment_var. By default, this is minimum() (assuming that column values in treatment_var are numerical).

Example

using MechanisticEffectModels.EcotoxSystems, MechanisticEffectModels.Utils

simfunct(x) = @replicates EcotoxSystems.simulator(x) 10

sim = exposure(simfunct, Params(), [0., 100., 200.]) |>
x -> relative_response(x, [:S, :H, :R]) # -> data frame will contain columns y_S, y_H, y_R for control-normalized values

EcotoxSystems.replicates — Method

replicates(simulator::Function, defaultparams::ComponentVector, nreps::Int64; kwargs...)

Perform replicated runs of simulator with parameters defaultparams (simulator(defaultparams) has to be a valid function call). Analogous to @replicates, but a bit more flexible.

EcotoxSystems.sig — Method

sig(
    x::Real, 
    x_thr::Real,
    y_left::Real, 
    y_right::Real; 
    beta::Real = 30
    )::Real

Sigmoid switch function. Used to replace simple if-statements with a continuous function in ODE models.

y_left and y_right are the function values left and right of the threshold x_thr.

EcotoxSystems.softNEC2GUTS — Method

Transformation of softNECpos for increasing relationships with domain (0,Inf).

EcotoxSystems.softNEC2neg — Method

Transformation of softNEC2pos for decreasing relationships with domain (0,1).

EcotoxSystems.softNEC2pos — Method

Linear positive relationship with soft threshold (NEC) and domain (1,Inf).

EcotoxSystems.texposure — Method

Threaded version of exposure().

EcotoxSystems.treplicates — Method

treplicates(
    simulator::Function, 
    defaultparams::ComponentVector, 
    nreps::Int64; 
    kwargs...)

Multi-threaded version of replicates.

Only useful if Julia has been started with multiple threads.

To check the number of threads, run using Base.Threads; Threads.nthreads().

In VSCode, you can use the entry "julia.NumThreads" in settings.json to set the default number of threads (searching for "julia threads" in the preferences will lead you there).

Check the Multi-threading documentation for more information.

EcotoxSystems.@replicates — Macro

@replicates(simcall::Expr, nreps::Int64)

Perform replicated runs of simcall, where simcall is a call to a simulator function.

Example:

    spc = SpeciesParams(Z = Truncated(Normal(1, 0.1), 0, Inf)) # initialize default parameters with variable zoom factor
    sim = @replicates MechanisticEffectModels.simulator(Params(spc = spc))) 10 # execute replicated runs to simulator

In this case, sim will contain the output of 10 replicated simulations. For each replicate, the zoom factor is sampled from a truncated Normal distribution. sim contains an additional column replicate.