EcotoxSystems.jl

Quickstart

To get acquainted with using EcotoxSystems.jl, you can use the default model and default parameters. <br>

Let's start with simulating individual life-history based on the default parameters:

p = deepcopy(EcotoxSystems.defaultparams)
p.glb.t_max = 21.
sim = EcotoxSystems.ODE_simulator(p)
first(sim, 5)

Here, we first create a copy of the default parameters. These contain two components:

glb for global parameters. This includes simulation settings like the maximum simulated time-span, and forcings such as the food input rate dX_in.

spc for species-specific parameters. For the default model, these are DEB-TKTD and some auxiliary parameters.

The second line,

p.glb.t_max = 21.

adjusts the simulated time-span. <br> With

sim = EcotoxSystems.ODE_simulator(p)

, we then simulate the model. <br> ODE_simulator is the central function to execute models which are purely ODE-based. <br> The output `sim is a DataFrame containing all state variables over time. <br> Storing the solution in a DataFrame can be convenient, but, for simple models, consumes a considerable proportion of computation time. It is possible to skip the conversion and retrieve the ODEsolution instead by setting returntype=EcoxSystems.odesol. <br>

Instead of simulating the life-history of a single individual, we can also simulate population dynamics. To do so for the default model, we need to adjust the global parameters, set plausible value for the starvation parameters (which were irrelevant before), and call IBM_simulator:

using Distributions, ProgressMeter

p = deepcopy(EcotoxSystems.defaultparams)

# adjusting global parameters

p.glb.t_max = 365. # simulated timespan [d]
p.glb.V_patch = 0.5 # simulated volume [L]
p.glb.dX_in = 10_000 # provide more food 

# adjusting species-level parameters

p.spc.Z = Truncated(Normal(1, 0.1), 0, Inf) # induce individual variability 
p.spc.W_S_rel_crit = 0.66
p.spc.h_S = -log(0.5)

sims = @replicates EcotoxSystems.IBM_simulator(p, showinfo = Inf) 10 # run replicated simulations

To deviate from the default model, there are a number of components to tweak in order to fully specify the system:

init_global_statevars: A function that initializes global state variables.

global_ode!: ODE-portion of the dynamics of global states.

global_rules!: Rule-based portion of the dynamics of global states.

init_individual_statevars: A function that initializes individual-level state variables.

individual_ode!: ODE-portion of the dynamics of individual-level states.

individual_rules!: Rule-based portion of the dynamics of individual-level states.

gen_ind_params: A function that translates species-specicif parameters spc into individual-level parameters ind. The default is generate_individual_params and is generic, apart from assuming that spc contains an entry for Z and propagate_zoom.

API

EcotoxSystems.IBM_simulator — Method

IBM_simulator(
    p::ComponentVector; 
    global_ode! = default_global_ODE!,
    global_rules! = default_global_rules!,
    init_global_statevars = initialize_global_statevars,
    individual_ode! = default_individual_ODE!,
    individual_rules! = default_individual_rules!,
    init_individual_statevars = initialize_individual_statevars,
    dt = 1/24, 
    saveat = 1,
    record_individuals = true,
    showinfo::Number = Inf
)

Simulate the individual-based version of the default model.

using EcotoxSystems
p = DEB.params()
sim = DEB.IBMsimulator(p)

For explanation of arguments, see IndividualBasedModel.

EcotoxSystems.ODE_simulator — Method

ODE_simulator(
    p::ComponentVector; 
    alg = Tsit5(),
    saveat = 1,
    reltol = 1e-6,
    model = default_ODE!,
    statevars_init = initialize_statevars,
    gen_ind_params = generate_individual_params,
    param_links::Union{Nothing,NamedTuple} = nothing,  
    callbacks = DEBODE_callbacks,
    returntype::ReturnType = dataframe,
    kwargs...
)

Run the model as ODE system. This function is essentially a wrapper around OrdinaryDiffEq.solve with additional input and output processing.

args:

p: Parameters, given as component vector.

At least two components are given: glb for global parameters, spc for species-level parameters. glb as to contain an entry t_max for the maximum simulation time. spc has to contain entries Z and propagate_zoom for the zoom factor and the parameters which are affected by the zoom factor.

kwargs:

The following kwargs are used internally by OrdinaryDiffEq.solve. See OrdinaryDiffEq documentation for more information.

alg: ODE solving algorithm to use..

saveat: Interval or time-points at which to save the solution. Default is 1.

reltol: Relative tolerance of ODE solution, default is 1e-6.

model: ODE system to solve.

callbacks: A callback set, i.e. pairs of conditions and events. Used to handle discontinuities.

In addition we have some kwargs that are used to further process inputs and outputs:

statevars_init: Function with signature statevars_init(p) that defines initial state variables as component vector. Components typically match those in the parameter vector.

`genindparams: Function that converts species-level parameters to individual-level parameters, for example by replacing parameters which are given as distributions with a random sample from the distribution. The inputs and outputs of this function should contain all components.

returntype: Indicating of how to return the result. Currently allowed are dataframe (complete solution converted to a DataFrame) and odesol (the ODE solution object as returned by OrdinaryDiffEq). Default is dataframe.

Run a model as purely ODE-based system:

using EcotoxSystems
sim = DEB.ODE_simulator(EcotoxSystems.defaultparams(p))

EcotoxSystems.exposure — Method

exposure(
    simulator::Function, 
    p::ComponentVector, 
    Cmat::Union{Vector{R},Matrix{R}}
) where R <: Real

Simulate constant chemical exposure with over an arbitrary number of treatments and chemical stressors, defined in Cmat.

The columns of C_W are stressors, the rows are treatments. <br> If C_W is supplied as a Vector, it is assumed that the elements represent different levels of single-stressor exposure. <br> That means, to simulate a single-stressor experiment with three treatments, do

Cmat = [0., 1., 2,]

, internally creating a 1 x 3 matrix with exposure concentrations 0, 1 and 2.

A single treatment with multiple stressors can be defined as

Cmat = [0 1 2;]

Here, we would have three stressors with the simultaneous exposure concentrations 0,1,2. <br>

Defining four treatments for two stressors looks like this:

Cmat = [
    0.0 0.0; 
    0.0 0.5; 
    1.0 0.0; 
    0.5 1.0
    ]

This exposure matrix corresponds to a ray design with constant exposure ratios.

EcotoxSystems.groupedlineplot! — Method

groupedlineplot(x, y, g, q; estimator)

Version of lineplot with additional grouping variable g.

EcotoxSystems.groupedlineplot! — Method

groupedlineplot(x, y, g, q; estimator)

Version of lineplot with additional grouping variable g.

EcotoxSystems.groupedlineplot — Method

groupedlineplot(x, y, g, q; estimator)

Version of lineplot with additional grouping variable g.

EcotoxSystems.lineplot! — Method

lineplot(x, y, q; estimator)

Plot aggregated values of y over x. Similar to seaborn.lineplot. By default, the arithmetic mean of y for every value of x is plotted, with 5th to 95th percentiles as ribbon. The plotted percentile range can be controlled via the positional argument q, which is a two-element Tuple. The arithmetic mean can be replaced with another function via the estimator argument.

EcotoxSystems.lineplot! — Method

lineplot(x, y, q; estimator)

Plot aggregated values of y over x. Similar to seaborn.lineplot. By default, the arithmetic mean of y for every value of x is plotted, with 5th to 95th percentiles as ribbon. The plotted percentile range can be controlled via the positional argument q, which is a two-element Tuple. The arithmetic mean can be replaced with another function via the estimator argument.

EcotoxSystems.lineplot — Method

lineplot(x, y, q; estimator)

Plot aggregated values of y over x. Similar to seaborn.lineplot. By default, the arithmetic mean of y for every value of x is plotted, with 5th to 95th percentiles as ribbon. The plotted percentile range can be controlled via the positional argument q, which is a two-element Tuple. The arithmetic mean can be replaced with another function via the estimator argument.

EcotoxSystems.relative_response — Method

relative_response(
    sim::D, 
    response_vars::Vector{Symbol},
    treatment_var::Symbol; 
    groupby_vars::Vector{Symbol} = Symbol[],
    identify_control = minimum
    ) where D <: AbstractDataFrame

Calculate relative responses.

args:

sim::AbstractDataFrame: results

response_vars::Vector{Symbol}: response variables for which to calculate the relative responses

treatment_var::Symbol: Column indicating the treatment. Column values can be numerical or categorical, but identify_control kwarg has to be specified in the latter case

kwargs:

groupby_vars::Vector{Symbol}: relative response will be conditioned on these variables (e.g. time, separate experiments...). Empty by default.

identify_control: function used to identify reference values from treatment_var. By default, this is minimum() (assuming that column values in treatment_var are numerical).

Example

using EcotoxSytems; EcotoxSystems.DEBkiss
sim = DEBkiss.simulate_constant_exposure(p, :C_W1 => [0, 1, 3, 10])
rr = relative_response(sim, [:S, :R], :C_W1; groupby_vars = [:t]) # calculates relative response over time for S and R

EcotoxSystems.replicates — Method

replicates(simulator::Function, defaultparams::ComponentVector, nreps::Int64; kwargs...)

Perform replicated runs of simulator with parameters defaultparams (simulator(defaultparams) has to be a valid function call). Analogous to @replicates, but a bit more flexible.

EcotoxSystems.texposure — Method

Threaded version of exposure().

EcotoxSystems.treplicates — Method

treplicates(
    simulator::Function, 
    defaultparams::ComponentVector, 
    nreps::Int64; 
    kwargs...)

Multi-threaded version of replicates.

Only useful if Julia has been started with multiple threads.

To check the number of threads, run using Base.Threads; Threads.nthreads().

In VSCode, you can use the entry "julia.NumThreads" in settings.json to set the default number of threads (searching for "julia threads" in the preferences will lead you there).

Check the Multi-threading documentation for more information.

EcotoxSystems.@replicates — Macro

@replicates(simcall::Expr, nreps::Int64)

Perform replicated runs of simcall, where simcall is a call to a simulator function.

Example:

    spc = SpeciesParams(Z = Truncated(Normal(1, 0.1), 0, Inf)) # initialize default parameters with variable zoom factor
    sim = @replicates MechanisticEffectModels.simulator(Params(spc = spc))) 10 # execute replicated runs to simulator

In this case, sim will contain the output of 10 replicated simulations. For each replicate, the zoom factor is sampled from a truncated Normal distribution. sim contains an additional column replicate.