EcotoxSystems.jl

API

EcotoxSystems.ODE_simulator — Method

ODE_simulator(
    p::ComponentVector; 
    alg = Tsit5(),
    saveat = 1,
    reltol = 1e-6,
    model = default_ODE!,
    statevars_init = initialize_statevars,
    gen_ind_params = generate_individual_params,
    param_links::Union{Nothing,NamedTuple} = nothing,  
    callbacks = DEBODE_callbacks,
    returntype::ReturnType = dataframe,
    kwargs...
)

Run the model as ODE system. This function is essentially a wrapper around OrdinaryDiffEq.solve with additional input and output processing.

args:

p: Parameters, given as component vector.

At least two components are given: glb for global parameters, spc for species-level parameters. glb as to contain an entry t_max for the maximum simulation time. spc has to contain entries Z and propagate_zoom for the zoom factor and the parameters which are affected by the zoom factor.

kwargs:

The following kwargs are used internally by OrdinaryDiffEq.solve. See OrdinaryDiffEq documentation for more information.

alg: ODE solving algorithm to use..

saveat: Interval or time-points at which to save the solution. Default is 1.

reltol: Relative tolerance of ODE solution, default is 1e-6.

model: ODE system to solve.

callbacks: A callback set, i.e. pairs of conditions and events. Used to handle discontinuities.

In addition we have some kwargs that are used to further process inputs and outputs:

statevars_init: Function with signature statevars_init(p) that defines initial state variables as component vector. Components typically match those in the parameter vector.

`genindparams: Function that converts species-level parameters to individual-level parameters, for example by replacing parameters which are given as distributions with a random sample from the distribution. The inputs and outputs of this function should contain all components.

returntype: Indicating of how to return the result. Currently allowed are dataframe (complete solution converted to a DataFrame) and odesol (the ODE solution object as returned by OrdinaryDiffEq). Default is dataframe.

Run a model as purely ODE-based system:

using EcotoxSystems
sim = DEB.ODE_simulator(EcotoxSystems.defaultparams(p))

EcotoxSystems.groupedlineplot! — Method

groupedlineplot(x, y, g, q; estimator)

Version of lineplot with additional grouping variable g.

EcotoxSystems.groupedlineplot! — Method

groupedlineplot(x, y, g, q; estimator)

Version of lineplot with additional grouping variable g.

EcotoxSystems.groupedlineplot — Method

groupedlineplot(x, y, g, q; estimator)

Version of lineplot with additional grouping variable g.

EcotoxSystems.lineplot! — Method

lineplot(x, y, q; estimator)

Plot aggregated values of y over x. Similar to seaborn.lineplot. By default, the arithmetic mean of y for every value of x is plotted, with 5th to 95th percentiles as ribbon. The plotted percentile range can be controlled via the positional argument q, which is a two-element Tuple. The arithmetic mean can be replaced with another function via the estimator argument.

EcotoxSystems.lineplot! — Method

lineplot(x, y, q; estimator)

Plot aggregated values of y over x. Similar to seaborn.lineplot. By default, the arithmetic mean of y for every value of x is plotted, with 5th to 95th percentiles as ribbon. The plotted percentile range can be controlled via the positional argument q, which is a two-element Tuple. The arithmetic mean can be replaced with another function via the estimator argument.

EcotoxSystems.lineplot — Method

lineplot(x, y, q; estimator)

Plot aggregated values of y over x. Similar to seaborn.lineplot. By default, the arithmetic mean of y for every value of x is plotted, with 5th to 95th percentiles as ribbon. The plotted percentile range can be controlled via the positional argument q, which is a two-element Tuple. The arithmetic mean can be replaced with another function via the estimator argument.

EcotoxSystems.relative_response — Method

relative_response(
    sim::D, 
    response_vars::Vector{Symbol},
    treatment_var::Symbol; 
    groupby_vars::Vector{Symbol} = Symbol[],
    identify_control = minimum
    ) where D <: AbstractDataFrame

Calculate relative responses.

args:

sim::AbstractDataFrame: results

response_vars::Vector{Symbol}: response variables for which to calculate the relative responses

treatment_var::Symbol: Column indicating the treatment. Column values can be numerical or categorical, but identify_control kwarg has to be specified in the latter case

kwargs:

groupby_vars::Vector{Symbol}: relative response will be conditioned on these variables (e.g. time, separate experiments...). Empty by default.

identify_control: function used to identify reference values from treatment_var. By default, this is minimum() (assuming that column values in treatment_var are numerical).

Example

using EcotoxSytems; EcotoxSystems.DEBkiss
sim = DEBkiss.simulate_constant_exposure(p, :C_W1 => [0, 1, 3, 10])
rr = relative_response(sim, [:S, :R], :C_W1; groupby_vars = [:t]) # calculates relative response over time for S and R

EcotoxSystems.texposure — Method

Threaded version of exposure().